requirements.txt
fastapi==0.115.7
pydantic==2.10.6
python-jose==3.3.0
loguru==0.7.3



tortoise-orm==0.24.0
asyncpg==0.30.0
aerich==0.7.2


uvicorn==0.34.0
gunicorn==23.0.0



python-dotenv==1.0.1

python-multipart==0.0.20
pytest-asyncio==0.25.3

bcrypt==4.3.0
httpx==0.27.2
aioboto3==14.1.0
tenacity==9.0.0




# Broker
aio-pika==9.5.5

# Cash
fastapi-cache2==0.2.2
redis==5.2.1


pydantic-settings==2.9.1

# Monitoring
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-instrumentation-fastapi==0.52b1
prometheus_client==0.21.1
